+++ src/main/java/com/zycus/boot/repositories/DepartmentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.zycus.boot.entities.Department;
 
 @Repository
-public interface DepartmentRepository extends CrudRepository<Department,Long> {
+public interface DepartmentRepository extends CrudRepository<Department,Integer> {
 
 	@Query("SELECT COUNT(D) FROM Department D WHERE D.departmentName = :departmentName")
 	Integer countByName(@Param("departmentName") String departmentName);
